refactor: move docker unit tests inline, keep only Docker-spawning tests in tests/
