+++ tests/docker.rs
@@ -1,441 +1,51 @@
 mod common;
 
 use capsule::docker::{
-    build_base_image, build_derived_image, build_docker_args, derived_image_name,
-    detect_compose_network, make_mcp_config, run_iteration, RunConfig, DOCKERFILE,
-    STREAM_DISPLAY_JQ,
+    build_base_image, build_derived_image, detect_compose_network, run_iteration, RunConfig,
 };
 use common::requires_docker;
 use serial_test::serial;
 use std::sync::{Arc, Mutex};
 
 #[test]
-fn embedded_dockerfile_is_non_empty() {
-    assert!(
-        !DOCKERFILE.is_empty(),
-        "embedded Dockerfile must not be empty"
-    );
-    assert!(
-        DOCKERFILE.contains("FROM archlinux"),
-        "Dockerfile must start from archlinux base"
-    );
-}
-
-#[test]
-fn embedded_stream_display_jq_is_non_empty() {
-    assert!(
-        !STREAM_DISPLAY_JQ.is_empty(),
-        "embedded stream_display.jq must not be empty"
-    );
-    assert!(
-        STREAM_DISPLAY_JQ.contains("fromjson"),
-        "jq filter must contain fromjson"
-    );
-}
-
-#[test]
-fn make_mcp_config_contains_binary_path_and_mcp_serve() {
-    let bin = std::path::Path::new("/usr/local/bin/capsule");
-    let cfg = make_mcp_config(bin);
-    let v: serde_json::Value = serde_json::from_str(&cfg).expect("valid JSON");
-    assert_eq!(
-        v["mcpServers"]["capsule"]["command"],
-        "/usr/local/bin/capsule"
-    );
-    assert_eq!(v["mcpServers"]["capsule"]["args"][0], "mcp-serve");
-}
-
-#[test]
-fn prompt_mount_is_not_read_only() {
+fn detect_compose_network_returns_none_when_no_project() {
     let dir = tempfile::tempdir().expect("temp dir");
-    let prompt_file = tempfile::NamedTempFile::new().unwrap();
-    let cfg = RunConfig {
-        pwd: dir.path().to_path_buf(),
-        ..RunConfig::default()
-    };
-    let args = build_docker_args(&cfg, prompt_file.path(), "capsule-test");
-    let prompt_arg = args.iter().find(|a| a.contains("prompt.txt")).unwrap();
+    let result = detect_compose_network(dir.path());
     assert!(
-        !prompt_arg.ends_with(":ro"),
-        "prompt.txt must not be mounted read-only so before-each.sh can mutate it: {prompt_arg}"
+        result.is_none(),
+        "expected None when no compose project running at temp dir"
     );
 }
 
 #[test]
-fn workspace_mounted_at_host_path_not_slash_workspace() {
+#[requires_docker]
+fn detect_compose_network_returns_network_for_running_project() {
     let dir = tempfile::tempdir().expect("temp dir");
-    let prompt_file = tempfile::NamedTempFile::new().unwrap();
-    let cfg = RunConfig {
-        pwd: dir.path().to_path_buf(),
-        ..RunConfig::default()
-    };
-    let args = build_docker_args(&cfg, prompt_file.path(), "capsule-test");
-    let joined = args.join(" ");
-    let pwd_str = dir.path().to_string_lossy();
-    assert!(
-        joined.contains(&format!("-v={pwd_str}:{pwd_str}")),
-        "workspace must be mounted at host path, not /workspace: {joined}"
-    );
-    assert!(
-        !joined.contains(":/workspace"),
-        "must not mount workspace at /workspace: {joined}"
-    );
-}
-
-#[test]
-fn workdir_set_to_host_path() {
-    let dir = tempfile::tempdir().expect("temp dir");
-    let prompt_file = tempfile::NamedTempFile::new().unwrap();
-    let cfg = RunConfig {
-        pwd: dir.path().to_path_buf(),
-        ..RunConfig::default()
-    };
-    let args = build_docker_args(&cfg, prompt_file.path(), "capsule-test");
-    let joined = args.join(" ");
-    let pwd_str = dir.path().to_string_lossy();
-    assert!(
-        joined.contains(&format!("--workdir={pwd_str}")),
-        "expected --workdir set to host path in args: {joined}"
-    );
-}
-
-#[test]
-fn capsule_workspace_env_var_set_to_host_path() {
-    let dir = tempfile::tempdir().expect("temp dir");
-    let prompt_file = tempfile::NamedTempFile::new().unwrap();
-    let cfg = RunConfig {
-        pwd: dir.path().to_path_buf(),
-        ..RunConfig::default()
-    };
-    let args = build_docker_args(&cfg, prompt_file.path(), "capsule-test");
-    let joined = args.join(" ");
-    let pwd_str = dir.path().to_string_lossy();
-    assert!(
-        joined.contains(&format!("-e=CAPSULE_WORKSPACE={pwd_str}")),
-        "expected -e=CAPSULE_WORKSPACE=<host-path> in args: {joined}"
-    );
-}
-
-#[test]
-fn env_file_arg_present_when_file_exists() {
-    let dir = tempfile::tempdir().expect("temp dir");
-    std::fs::write(dir.path().join(".env"), "FOO=bar\n").unwrap();
-
-    let prompt_file = tempfile::NamedTempFile::new().unwrap();
-    let cfg = RunConfig {
-        pwd: dir.path().to_path_buf(),
-        env_file: Some(dir.path().join(".env")),
-        ..RunConfig::default()
-    };
-    let args = build_docker_args(&cfg, prompt_file.path(), "capsule-test");
-    let joined = args.join(" ");
-    assert!(
-        joined.contains("--env-file"),
-        "expected --env-file in args: {joined}"
-    );
-    assert!(
-        joined.contains(".env"),
-        "expected .env path in args: {joined}"
-    );
-}
-
-#[test]
-fn env_file_arg_absent_when_no_file() {
-    let dir = tempfile::tempdir().expect("temp dir");
-
-    let prompt_file = tempfile::NamedTempFile::new().unwrap();
-    let cfg = RunConfig {
-        pwd: dir.path().to_path_buf(),
-        ..RunConfig::default()
-    };
-    let args = build_docker_args(&cfg, prompt_file.path(), "capsule-test");
-    let joined = args.join(" ");
-    assert!(
-        !joined.contains("--env-file"),
-        "expected no --env-file when env_file is None: {joined}"
-    );
-}
-
-#[test]
-fn gh_token_env_file_passed_when_present() {
-    let dir = tempfile::tempdir().expect("temp dir");
-    let token_file = dir.path().join("gh-token.env");
-    std::fs::write(&token_file, "GH_TOKEN=ghs_testtoken\n").unwrap();
-
-    let prompt_file = tempfile::NamedTempFile::new().unwrap();
-    let cfg = RunConfig {
-        pwd: dir.path().to_path_buf(),
-        gh_token_env_file: Some(token_file.clone()),
-        ..RunConfig::default()
-    };
-    let args = build_docker_args(&cfg, prompt_file.path(), "capsule-test");
-    let joined = args.join(" ");
-    assert!(
-        joined.contains("--env-file"),
-        "expected --env-file for gh token: {joined}"
-    );
-    assert!(
-        joined.contains("gh-token.env"),
-        "expected token file path in args: {joined}"
-    );
-}
-
-#[test]
-fn gh_token_not_in_docker_args_when_env_file_none() {
-    let dir = tempfile::tempdir().expect("temp dir");
-
-    let prompt_file = tempfile::NamedTempFile::new().unwrap();
-    let cfg = RunConfig {
-        pwd: dir.path().to_path_buf(),
-        ..RunConfig::default()
-    };
-    let args = build_docker_args(&cfg, prompt_file.path(), "capsule-test");
-    let joined = args.join(" ");
-    assert!(
-        !joined.contains("GH_TOKEN"),
-        "token must not appear in docker args: {joined}"
-    );
-}
-
-#[test]
-fn gh_token_never_appears_inline_in_docker_args() {
-    // Even if a token string is known, it must not show up in the arg list directly.
-    // The only valid path is via --env-file.
-    let dir = tempfile::tempdir().expect("temp dir");
-    let token_file = dir.path().join("gh-token.env");
-    std::fs::write(&token_file, "GH_TOKEN=ghs_secret\n").unwrap();
-
-    let prompt_file = tempfile::NamedTempFile::new().unwrap();
-    let cfg = RunConfig {
-        pwd: dir.path().to_path_buf(),
-        gh_token_env_file: Some(token_file),
-        ..RunConfig::default()
-    };
-    let args = build_docker_args(&cfg, prompt_file.path(), "capsule-test");
-    for arg in &args {
-        assert!(
-            !arg.contains("ghs_secret"),
-            "token value must not appear inline in docker arg: {arg}"
-        );
-    }
-}
-
-#[test]
-fn git_config_mounted_readonly_when_present() {
-    let dir = tempfile::tempdir().expect("temp dir");
-    let git_dir = dir.path().join(".git");
-    std::fs::create_dir(&git_dir).unwrap();
+    let compose_file = dir.path().join("docker-compose.yml");
     std::fs::write(
-        git_dir.join("config"),
-        "[core]\n\trepositoryformatversion = 0\n",
+        &compose_file,
+        "services:\n  web:\n    image: busybox\n    command: sleep 30\n",
     )
     .unwrap();
 
-    let prompt_file = tempfile::NamedTempFile::new().unwrap();
-    let cfg = RunConfig {
-        pwd: dir.path().to_path_buf(),
-        ..RunConfig::default()
-    };
-    let args = build_docker_args(&cfg, prompt_file.path(), "capsule-test");
-    let joined = args.join(" ");
-    let pwd_str = dir.path().to_string_lossy();
-    assert!(
-        joined.contains(&format!(".git/config:{pwd_str}/.git/config:ro")),
-        "expected read-only git config mount at host path in args: {joined}"
-    );
-}
+    std::process::Command::new("docker")
+        .args(["compose", "-f", &compose_file.to_string_lossy(), "up", "-d"])
+        .current_dir(dir.path())
+        .output()
+        .expect("docker compose up should run");
 
-#[test]
-fn git_config_mount_absent_when_no_git_dir() {
-    let dir = tempfile::tempdir().expect("temp dir");
+    std::thread::sleep(std::time::Duration::from_secs(2));
 
-    let prompt_file = tempfile::NamedTempFile::new().unwrap();
-    let cfg = RunConfig {
-        pwd: dir.path().to_path_buf(),
-        ..RunConfig::default()
-    };
-    let args = build_docker_args(&cfg, prompt_file.path(), "capsule-test");
-    let joined = args.join(" ");
+    let result = detect_compose_network(dir.path());
     assert!(
-        !joined.contains(".git/config"),
-        "expected no git config mount when .git/config absent: {joined}"
+        result.is_some(),
+        "expected a network name for running compose project"
     );
-}
 
-#[test]
-fn git_identity_env_vars_present_in_docker_args() {
-    let dir = tempfile::tempdir().expect("temp dir");
-    let prompt_file = tempfile::NamedTempFile::new().unwrap();
-    let cfg = RunConfig {
-        pwd: dir.path().to_path_buf(),
-        git_author_name: "Bob Builder".to_string(),
-        git_author_email: "bob@example.com".to_string(),
-        ..RunConfig::default()
-    };
-    let args = build_docker_args(&cfg, prompt_file.path(), "capsule-test");
-    let joined = args.join(" ");
-    assert!(
-        joined.contains("GIT_AUTHOR_NAME=Bob Builder"),
-        "expected GIT_AUTHOR_NAME in args: {joined}"
-    );
-    assert!(
-        joined.contains("GIT_AUTHOR_EMAIL=bob@example.com"),
-        "expected GIT_AUTHOR_EMAIL in args: {joined}"
-    );
-    assert!(
-        joined.contains("GIT_COMMITTER_NAME=Bob Builder"),
-        "expected GIT_COMMITTER_NAME in args: {joined}"
-    );
-    assert!(
-        joined.contains("GIT_COMMITTER_EMAIL=bob@example.com"),
-        "expected GIT_COMMITTER_EMAIL in args: {joined}"
-    );
-}
-
-#[test]
-fn git_identity_env_vars_present_when_empty() {
-    let dir = tempfile::tempdir().expect("temp dir");
-    let prompt_file = tempfile::NamedTempFile::new().unwrap();
-    let cfg = RunConfig {
-        pwd: dir.path().to_path_buf(),
-        ..RunConfig::default()
-    };
-    let args = build_docker_args(&cfg, prompt_file.path(), "capsule-test");
-    let joined = args.join(" ");
-    // Even empty values should be passed so the entrypoint can fall back.
-    assert!(
-        joined.contains("GIT_AUTHOR_NAME="),
-        "expected GIT_AUTHOR_NAME= in args: {joined}"
-    );
-    assert!(
-        joined.contains("GIT_AUTHOR_EMAIL="),
-        "expected GIT_AUTHOR_EMAIL= in args: {joined}"
-    );
-}
-
-#[test]
-fn before_each_mounted_when_path_provided() {
-    let dir = tempfile::tempdir().expect("temp dir");
-    let before_each = dir.path().join("before-each.sh");
-    std::fs::write(&before_each, "#!/bin/sh\n").unwrap();
-
-    let prompt_file = tempfile::NamedTempFile::new().unwrap();
-    let cfg = RunConfig {
-        pwd: dir.path().to_path_buf(),
-        before_each_path: Some(before_each.clone()),
-        ..RunConfig::default()
-    };
-    let args = build_docker_args(&cfg, prompt_file.path(), "capsule-test");
-    let joined = args.join(" ");
-    assert!(
-        joined.contains("/home/claude/before-each.sh:ro"),
-        "expected before-each.sh mount in args: {joined}"
-    );
-    assert!(
-        joined.contains(before_each.to_string_lossy().as_ref()),
-        "expected host path in mount: {joined}"
-    );
-}
-
-#[test]
-fn before_each_not_mounted_when_absent() {
-    let dir = tempfile::tempdir().expect("temp dir");
-    let prompt_file = tempfile::NamedTempFile::new().unwrap();
-    let cfg = RunConfig {
-        pwd: dir.path().to_path_buf(),
-        ..RunConfig::default()
-    };
-    let args = build_docker_args(&cfg, prompt_file.path(), "capsule-test");
-    let joined = args.join(" ");
-    assert!(
-        !joined.contains("before-each.sh"),
-        "before-each.sh must not appear in args when path is None: {joined}"
-    );
-}
-
-#[test]
-fn model_arg_present_when_model_set() {
-    let dir = tempfile::tempdir().expect("temp dir");
-    let prompt_file = tempfile::NamedTempFile::new().unwrap();
-    let cfg = RunConfig {
-        pwd: dir.path().to_path_buf(),
-        model: Some("claude-opus-4-6".to_string()),
-        ..RunConfig::default()
-    };
-    let args = build_docker_args(&cfg, prompt_file.path(), "capsule-test");
-    let joined = args.join(" ");
-    assert!(
-        joined.contains("-e=CAPSULE_MODEL=claude-opus-4-6"),
-        "expected -e=CAPSULE_MODEL=claude-opus-4-6 in args: {joined}"
-    );
-}
-
-#[test]
-fn model_arg_absent_when_no_model() {
-    let dir = tempfile::tempdir().expect("temp dir");
-    let prompt_file = tempfile::NamedTempFile::new().unwrap();
-    let cfg = RunConfig {
-        pwd: dir.path().to_path_buf(),
-        ..RunConfig::default()
-    };
-    let args = build_docker_args(&cfg, prompt_file.path(), "capsule-test");
-    let joined = args.join(" ");
-    assert!(
-        !joined.contains("CAPSULE_MODEL"),
-        "CAPSULE_MODEL must not appear in args when model is None: {joined}"
-    );
-}
-
-#[test]
-fn verbose_flag_not_added_to_docker_args() {
-    // verbose is host-side behavior; it must not add extra docker flags.
-    let dir = tempfile::tempdir().expect("temp dir");
-    let prompt_file = tempfile::NamedTempFile::new().unwrap();
-    let cfg_verbose = RunConfig {
-        pwd: dir.path().to_path_buf(),
-        verbose: true,
-        ..RunConfig::default()
-    };
-    let cfg_quiet = RunConfig {
-        pwd: dir.path().to_path_buf(),
-        ..RunConfig::default()
-    };
-    let args_verbose = build_docker_args(&cfg_verbose, prompt_file.path(), "capsule-test");
-    let args_quiet = build_docker_args(&cfg_quiet, prompt_file.path(), "capsule-test");
-    assert_eq!(
-        args_verbose, args_quiet,
-        "verbose flag must not alter docker args"
-    );
-}
-
-#[test]
-fn container_name_present_in_docker_args() {
-    let dir = tempfile::tempdir().expect("temp dir");
-    let prompt_file = tempfile::NamedTempFile::new().unwrap();
-    let cfg = RunConfig {
-        pwd: dir.path().to_path_buf(),
-        ..RunConfig::default()
-    };
-    let args = build_docker_args(&cfg, prompt_file.path(), "capsule-run-12345-1");
-    let joined = args.join(" ");
-    assert!(
-        joined.contains("--name capsule-run-12345-1"),
-        "expected --name in args: {joined}"
-    );
-}
-
-#[test]
-fn container_name_for_has_expected_format() {
-    use capsule::docker::container_name_for;
-    let name = container_name_for(3);
-    assert!(
-        name.starts_with("capsule-run-"),
-        "name should start with capsule-run-: {name}"
-    );
-    assert!(
-        name.ends_with("-3"),
-        "name should end with iteration number: {name}"
-    );
+    let _ = std::process::Command::new("docker")
+        .args(["compose", "-f", &compose_file.to_string_lossy(), "down"])
+        .current_dir(dir.path())
+        .output();
 }
 
 #[test]
@@ -445,7 +55,6 @@ fn run_iteration_with_model_passes_capsule_model_to_container() {
     let workdir = tempfile::tempdir().expect("temp workdir");
     let output_file = workdir.path().join("model_output.txt");
 
-    // Entrypoint: write $CAPSULE_MODEL to $CAPSULE_WORKSPACE/model_output.txt then exit 0.
     let dockerfile =
         "FROM busybox\nENTRYPOINT [\"sh\", \"-c\", \"echo \\\"$CAPSULE_MODEL\\\" > \\\"$CAPSULE_WORKSPACE/model_output.txt\\\"; exit 0\"]\n";
     let mut child = std::process::Command::new("docker")
@@ -485,135 +94,9 @@ fn run_iteration_with_model_passes_capsule_model_to_container() {
         "container should receive CAPSULE_MODEL=claude-opus-4-6, got: {written:?}"
     );
 
-    // Cleanup
     let _ = std::process::Command::new("docker")
         .args(["rmi", "-f", "capsule-test-model"])
         .output();
-}
-
-#[test]
-fn compose_network_arg_present_when_set() {
-    let dir = tempfile::tempdir().expect("temp dir");
-    let prompt_file = tempfile::NamedTempFile::new().unwrap();
-    let cfg = RunConfig {
-        pwd: dir.path().to_path_buf(),
-        compose_network: Some("myproject_default".to_string()),
-        ..RunConfig::default()
-    };
-    let args = build_docker_args(&cfg, prompt_file.path(), "capsule-test");
-    let joined = args.join(" ");
-    assert!(
-        joined.contains("--network myproject_default"),
-        "expected --network in args: {joined}"
-    );
-}
-
-#[test]
-fn compose_network_arg_absent_when_none() {
-    let dir = tempfile::tempdir().expect("temp dir");
-    let prompt_file = tempfile::NamedTempFile::new().unwrap();
-    let cfg = RunConfig {
-        pwd: dir.path().to_path_buf(),
-        ..RunConfig::default()
-    };
-    let args = build_docker_args(&cfg, prompt_file.path(), "capsule-test");
-    let joined = args.join(" ");
-    assert!(
-        !joined.contains("--network"),
-        "expected no --network when compose_network is None: {joined}"
-    );
-}
-
-#[test]
-fn claude_dir_mounted_at_home_claude_dot_claude() {
-    let dir = tempfile::tempdir().expect("temp dir");
-    let claude_dir = tempfile::tempdir().expect("claude temp dir");
-    let prompt_file = tempfile::NamedTempFile::new().unwrap();
-    let cfg = RunConfig {
-        pwd: dir.path().to_path_buf(),
-        claude_dir: claude_dir.path().to_path_buf(),
-        ..RunConfig::default()
-    };
-    let args = build_docker_args(&cfg, prompt_file.path(), "capsule-test");
-    let joined = args.join(" ");
-    assert!(
-        joined.contains(":/home/claude/.claude"),
-        "expected ~/.claude mount in args: {joined}"
-    );
-    assert!(
-        joined.contains(claude_dir.path().to_string_lossy().as_ref()),
-        "expected host claude_dir path in mount: {joined}"
-    );
-}
-
-#[test]
-fn detect_compose_network_returns_none_when_no_project() {
-    let dir = tempfile::tempdir().expect("temp dir");
-    let result = detect_compose_network(dir.path());
-    assert!(
-        result.is_none(),
-        "expected None when no compose project running at temp dir"
-    );
-}
-
-#[test]
-#[requires_docker]
-fn detect_compose_network_returns_network_for_running_project() {
-    let dir = tempfile::tempdir().expect("temp dir");
-    let compose_file = dir.path().join("docker-compose.yml");
-    std::fs::write(
-        &compose_file,
-        "services:\n  web:\n    image: busybox\n    command: sleep 30\n",
-    )
-    .unwrap();
-
-    // Start compose project.
-    std::process::Command::new("docker")
-        .args(["compose", "-f", &compose_file.to_string_lossy(), "up", "-d"])
-        .current_dir(dir.path())
-        .output()
-        .expect("docker compose up should run");
-
-    // Give it a moment to start.
-    std::thread::sleep(std::time::Duration::from_secs(2));
-
-    let result = detect_compose_network(dir.path());
-    assert!(
-        result.is_some(),
-        "expected a network name for running compose project"
-    );
-
-    // Cleanup
-    let _ = std::process::Command::new("docker")
-        .args(["compose", "-f", &compose_file.to_string_lossy(), "down"])
-        .current_dir(dir.path())
-        .output();
-}
-
-#[test]
-fn derived_image_name_uses_basename_of_pwd() {
-    let dir = tempfile::tempdir().expect("temp dir");
-    let project_dir = dir.path().join("my-project");
-    std::fs::create_dir(&project_dir).unwrap();
-    let name = derived_image_name(&project_dir);
-    assert_eq!(name, "capsule-my-project");
-}
-
-#[test]
-fn derived_image_name_handles_root_or_unnamed() {
-    // If basename is empty (unlikely in practice), should not panic.
-    let name = derived_image_name(std::path::Path::new("/"));
-    assert!(name.starts_with("capsule-"), "name={name}");
-}
-
-#[test]
-fn build_derived_image_returns_none_when_no_dockerfile() {
-    let capsule_dir = tempfile::tempdir().expect("temp dir");
-    let pwd = tempfile::tempdir().expect("temp dir");
-    // No Dockerfile in capsule_dir → returns None without touching Docker.
-    let result = build_derived_image(capsule_dir.path(), pwd.path(), false)
-        .expect("should not error when Dockerfile absent");
-    assert!(result.is_none(), "expected None when no Dockerfile");
 }
 
 #[test]
@@ -624,8 +107,6 @@ fn build_derived_image_builds_and_returns_image_name() {
     let pwd = base.path().join("myproject");
     std::fs::create_dir(&pwd).unwrap();
 
-    // Write a minimal Dockerfile. Uses busybox so the test doesn't require the
-    // capsule base image to be pre-built.
     std::fs::write(
         capsule_dir.path().join("Dockerfile"),
         "FROM busybox\nRUN echo derived\n",
@@ -641,7 +122,6 @@ fn build_derived_image_builds_and_returns_image_name() {
         "derived image name should start with capsule-: {name}"
     );
 
-    // Cleanup: remove the derived image.
     let _ = std::process::Command::new("docker")
         .args(["rmi", "-f", &name])
         .output();
@@ -661,12 +141,9 @@ fn build_derived_image_skips_build_when_image_exists_and_no_rebuild() {
     )
     .unwrap();
 
-    // First build.
     let name = build_derived_image(capsule_dir.path(), &pwd, false)
         .unwrap()
         .unwrap();
-
-    // Second call with rebuild=false — should succeed without rebuilding.
     let name2 = build_derived_image(capsule_dir.path(), &pwd, false)
         .unwrap()
         .unwrap();
@@ -799,17 +276,14 @@ fn build_derived_image_rebuilds_when_dockerfile_changes() {
         .unwrap()
         .unwrap();
 
-    // Record image ID from first build.
     let id1_out = std::process::Command::new("docker")
         .args(["image", "inspect", "--format", "{{.Id}}", &name])
         .output()
         .unwrap();
     let id1 = String::from_utf8(id1_out.stdout).unwrap().trim().to_owned();
 
-    // Change the Dockerfile content.
     std::fs::write(&dockerfile_path, "FROM busybox\nRUN echo version2\n").unwrap();
 
-    // Should auto-rebuild without --rebuild flag.
     let name2 = build_derived_image(capsule_dir.path(), &pwd, false)
         .unwrap()
         .unwrap();
@@ -859,7 +333,6 @@ fn mcp_serve_handles_initialize_and_submit_verdict_in_container() {
     let stdout = child.stdout.take().unwrap();
     let mut reader = BufReader::new(stdout);
 
-    // initialize
     writeln!(
         stdin,
         r#"{{"jsonrpc":"2.0","id":0,"method":"initialize","params":{{"protocolVersion":"2024-11-05","capabilities":{{}},"clientInfo":{{"name":"test","version":"0.1"}}}}}}"#
